requirements.txt
opencv-python
ultralytics
coppeliasim-zmqremoteapi-client
pygame
SpeechRecognition
pyaudio
llama-cpp-python
requests
numpy